requirements.txt
amqp==2.1.4
Babel==2.3.4
billiard==3.5.0.2
celery==4.0.1
falcon==1.1.0
imagesize==0.7.1
Jinja2==2.8
kombu==4.0.1
MarkupSafe==0.23
py==1.4.31
Pygments==2.1.3
python-mimeparse==1.6.0
pytz==2016.10
requests==2.12.4
six==1.10.0
dnspython==1.15.0
snowballstemmer==1.2.1
vine==1.1.3
PyYAML==3.12
